requirements.txt
streamlit==1.41.0
groq==0.13.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0